feat: Add login functionality and dashboard layout
Adds login functionality to the application by creating the `LoginController` with an `index` method to show the login view and a `store` method to validate and authenticate user login credentials. On unsuccessful login attempts, the user is redirected back with an error message. After successful login, the user is redirected to the 'posts.index' route.

Also introduces a new dashboard layout by creating the `dashboard.blade.php` view file in the `resources/views/layouts` directory.